Split into smaller files.
No renaming done yet. The renaming should probably be mostly about adding an MM prefix.
